requirements.txt
backports.tarfile==1.2.0
bitarray==3.5.2
flask==3.1.1
importlib-metadata==8.0.0
jaraco.collections==5.1.0
matplotlib==3.10.5
pandas==2.3.1
pip-chill==1.0.3
platformdirs==4.2.2
psutil==7.0.0
pycryptodome==3.23.0
tabulate==0.9.0
tomli==2.0.1
PyQt6>=6.9
qt-material>=2.17
